requirements.txt
django-qrcode~=0.3
flask~=3.0.3